+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All features and notable changes to this project will be documented in this file
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
+## [0.1.3] - 2023-03-10
+
+### Fixed
+- (udp) sending confirm only once
+- (udp) processing of reply message
+- (udp) bye message would not be send when received SIGINT
+
 ## [0.1.2] - 2023-03-09
 
 ### Added